main/xiaozhi-server/requirements.txt
pyyml==0.0.2
torch==2.7.1
silero_vad==5.1.2
websockets==14.2
opuslib_next==1.1.2
numpy==1.26.4
pydub==0.25.1
funasr==1.2.3
torchaudio==2.7.1
openai==1.61.0
google-generativeai==0.8.4
edge_tts==7.0.0
httpx==0.27.2
aiohttp==3.9.3
aiohttp_cors==0.7.0
ormsgpack==1.7.0
ruamel.yaml==0.18.10
loguru==0.7.3
requests==2.32.3
cozepy==0.12.0
mem0ai==0.1.62
bs4==0.0.2
modelscope==1.23.2
# sherpa_onnx==1.12.8
mcp==1.8.1
cnlunar==0.2.0
PySocks==1.7.1
dashscope==1.23.1
baidu-aip==4.16.13
chardet==5.2.0
aioconsole==0.8.1
markitdown==0.1.1
mcp-proxy==0.8.0
PyJWT==2.8.0
psutil==7.0.0
portalocker==2.10.1
Jinja2==3.1.6
future==0.17.1
langchain
langchain-community
langchain-qdrant
openpyxl==3.1.5
pandas==2.2.3
qdrant-client==1.13.2
redis==5.2.1
xlrd==2.0.1
sentence-transformers
langchain-huggingface